make the link with Customer
