+++ src/main/Customer.java
@@ -11,7 +11,31 @@ package main;
  */
 public class Customer {
     
-    int Id;
-    String name;
+    //private int Id;
+    private String name;
+    private String surname;
+
+    public Customer(String name, String surname) {
+        this.name = name;
+        this.surname = surname;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getSurname() {
+        return surname;
+    }
+
+    public void setSurname(String surname) {
+        this.surname = surname;
+    }
+    
+    
     
 }
